part 4

+++ part4/src/controllers/blog.js
@@ -1,30 +1,58 @@
 const express = require("express");
-const expressRouter = express.Router();
 const mongoBlog = require("../models/mongo");
+const mongoUsers = require("../models/mongousers");
+const jwt = require("jsonwebtoken");
+const expressRouter = express.Router();
+
 
 
 expressRouter.get('/', async (request, response) => {
-  const result = await mongoBlog.find({});   //Find all blogs, empty means all
+  const result = await mongoBlog.find({}).populate("user", {blogs: 0});   //Find all blogs, empty means all
 
-  response.json(result)
+  response.json(result.map(blog => blog))
 });
 
 expressRouter.post('/', async (request, response) => {
   if((request.body.title === undefined) || (request.body.url === undefined)){
     return response.status(400).send({error: "Bad Request"})
   }
+
+  const decodedtoken = jwt.decode(request.token, "test");
+
+  if(!(decodedtoken.username || token)){
+    return response.status(400).send({error: "missing or invalid token"})
+  }
+
+  const user = await mongoUsers.findOne({username: decodedtoken.username});
+
   if(request.body.likes === undefined){
     request.body.likes = 0
   }
 
-  const blog = new mongoBlog(request.body);
+  const newblog = {
+    title: request.body.title,
+    author: request.body.author,
+    url: request.body.url,
+    likes: request.body.likes,
+    user: user._id
+  };
 
+  const blog = new mongoBlog(newblog);
   const saved = await blog.save();
+  user.blogs = user.blogs.concat(saved._id);
+  await user.save();
 
   response.json(saved.toJSON())
 });
 
 expressRouter.delete('/:id', async (request, response) => {
+  const decodedtoken = jwt.decode(request.token, "test");
+
+  if(!decodedtoken.username){
+    return response.status(400).send({error: "missing or invalid token"})
+  }
+
+
   await mongoBlog.findByIdAndDelete(request.params.id)
     .catch(error => response.status(400).send({error: error.message}));
 
@@ -40,7 +68,7 @@ expressRouter.put('/:id', async (request, response) => {
     runValidators: true,
     context: "query"
   })
-    .catch(error => response.status(400).send({error: error.message}))
+    .catch(error => response.status(400).send({error: error.message}));
 
   response.status(200).json(updated.toJSON())
